+++ docs/versionhistory.rst
@@ -3,6 +3,11 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
+**UNRELEASED**
+
+- Implemented a compatibility fix for supporting direct access of ``anyio.*`` submodules
+  from the main package even when those submodules were not directly imported first
+
 **4.15.0**
 
 - Added support for the newer keyword-only arguments on ``anyio.Path`` methods to match